+++ src/lambdas/sqs/send-event-reminder/processor.ts
@@ -6,10 +6,6 @@ import {
   type ActionableEventReminderAttemptSentEvent
 } from '@model/app-events/ActionableEventReminderAttemptSentEvent';
 import {
-  actionableEventReminderInsufficientCreditNotSent,
-  type ActionableEventReminderInsufficientCreditNotSentEvent
-} from '@model/app-events/ActionableEventReminderInsufficientCreditNotSentEvent';
-import {
   demoReminderLimitReachedNotSent,
   type DemoReminderLimitReachedNotSentEvent
 } from '@model/app-events/DemoReminderLimitReachedNotSentEvent';
@@ -18,6 +14,10 @@ import {
   type DemoReminderToBeSentAttemptSentEvent
 } from '@model/app-events/DemoReminderToBeSentAttemptSentEvent';
 import type { DemoReminderToBeSentEvent } from '@model/app-events/DemoReminderToBeSentEvent';
+import {
+  insufficientCreditReminderNotSent,
+  type InsufficientCreditReminderNotSentEvent
+} from '@model/app-events/InsufficientCreditReminderNotSentEvent';
 import type { CreditServiceEndpointConfig, DemoReminderEndpointConfig } from '@model/Config';
 import type { VonageEndpointConfig } from '@model/vendor/vonage/config';
 import type { IdpName, UserId, Uuid } from '@notifycal/shared/types';
@@ -141,21 +141,21 @@ export default class Processor {
     event: ActionableEventFoundEvent | DemoReminderToBeSentEvent
   ): Promise<Uuid> {
     return match(result)
-      .with({ operationId: 'InsufficientCredits' }, (insufficientResult) => {
+      .with({ result: 'InsufficientCredits' }, (insufficientResult) => {
         logger.info('Message not sent due to insufficient credits', { result });
         return this.publishInsufficientCreditErrorEvent(
           event as ActionableEventFoundEvent,
           insufficientResult as CreditDeductionInsufficientCreditsError
         ).then(() => 'insufficient-credits' as Uuid);
       })
-      .with({ operationId: 'DemoCounterLimitReachedError' }, (demoLimitResult) => {
+      .with({ result: 'DemoCounterLimitReachedError' }, (demoLimitResult) => {
         logger.info('Demo reminder not sent due to demo limit reached', { result });
         return this.publishDemoLimitReachedErrorEvent(
           event as DemoReminderToBeSentEvent,
           demoLimitResult as DemoCounterLimitReachedError
         ).then(() => 'demo-limit-reached' as Uuid);
       })
-      .with({ operationId: 'BadRequestError' }, () => {
+      .with({ result: 'BadRequestError' }, () => {
         const operationType =
           event.eventType === 'ActionableEventFound'
             ? 'credit deduction'
@@ -166,7 +166,7 @@ export default class Processor {
           })
         );
       })
-      .with({ operationId: 'UnknownError' }, () => {
+      .with({ result: 'UnknownError' }, () => {
         const operationType =
           event.eventType === 'ActionableEventFound'
             ? 'credit deduction'
@@ -219,12 +219,9 @@ export default class Processor {
     logger.info(
       'Publishing an event indicating a message could not be sent due to user having insufficient credits'
     );
-    const insufficientCreditEvent = actionableEventReminderInsufficientCreditNotSent(
-      event,
-      creditError
-    );
+    const insufficientCreditEvent = insufficientCreditReminderNotSent(event, creditError);
     return this.snsService.safePublish<
-      ActionableEventReminderInsufficientCreditNotSentEvent | DemoReminderLimitReachedNotSentEvent
+      InsufficientCreditReminderNotSentEvent | DemoReminderLimitReachedNotSentEvent
     >(insufficientCreditEvent);
   }
 
@@ -237,7 +234,7 @@ export default class Processor {
     );
     const demoLimitEvent = demoReminderLimitReachedNotSent(event, demoLimitError);
     return this.snsService.safePublish<
-      ActionableEventReminderInsufficientCreditNotSentEvent | DemoReminderLimitReachedNotSentEvent
+      InsufficientCreditReminderNotSentEvent | DemoReminderLimitReachedNotSentEvent
     >(demoLimitEvent);
   }
 }