+++ src/testing/data/app-events.ts
@@ -16,6 +16,7 @@ import {
   type EventId,
   type IdpId,
   type RCSSenderId,
+  type TemplateId,
   type TimeZone,
   type UserId
 } from '@notifycal/shared/types';
@@ -44,7 +45,7 @@ export const userCalendarFetchedEvent: UserCalendarFetchedEvent = {
       name: 'Some Calendar Name' as CalendarName
     },
     template: {
-      id: templateMap['formal-en-01'].id,
+      id: templateMap['formal-en-01']?.id ?? ('formal-en-01' as TemplateId),
       fields: {
         business: {
           name: 'SomeBusinessName' as BusinessName,
@@ -195,5 +196,35 @@ export const auditTrailNoPhoneNumberForCalendarEventFoundEvent: AuditTrailStoreR
       timeZone: 'Europe/Madrid' as TimeZone
     }
   },
+  Origin: 'somewhere' as AuditTrailStoreRecordOrigin
+};
+
+export const lowCreditsDetectedEvent: AuditTrailStoreRecord = {
+  Data: {
+    originalEvent: validActionableEventEvent.data,
+    error: { message: 'Low credits detected' }
+  },
+  CorrelationId: '0de651ef-535e-4d2e-b9ff-7bf43f5aaaaa' as CorrelationId,
+  EventId: '0de651ef-535e-4d2e-b9ff-7bf43f5a01ac' as EventId,
+  UserId: '0de651ef-535e-4d2e-b9ff-7bf43f5a0000' as UserId,
+  Idp: 'google.com',
+  IdpId: '45346356356' as IdpId,
+  EventType: 'LowCreditsDetected',
+  HappenedAt: '2024-01-02T15:04:50Z' as DateTime,
+  Origin: 'somewhere' as AuditTrailStoreRecordOrigin
+};
+
+export const insufficientCreditsReminderNotSentEvent: AuditTrailStoreRecord = {
+  Data: {
+    originalEvent: validActionableEventEvent.data,
+    error: { message: 'Insufficient credit reminder not sent' }
+  },
+  CorrelationId: '0de651ef-535e-4d2e-b9ff-7bf43f5aaaab' as CorrelationId,
+  EventId: '0de651ef-535e-4d2e-b9ff-7bf43f5a01ab' as EventId,
+  UserId: '0de651ef-535e-4d2e-b9ff-7bf43f5a0001' as UserId,
+  Idp: 'google.com',
+  IdpId: '45346356357' as IdpId,
+  EventType: 'InsufficientCreditReminderNotSent',
+  HappenedAt: '2024-01-02T15:04:51Z' as DateTime,
   Origin: 'somewhere' as AuditTrailStoreRecordOrigin
 };