+++ src/testing/utils/config.ts
@@ -42,9 +42,9 @@ export const fakeIdpConfigs: IdpConfigs = {
   }
 };
 
-function setEnvVar(envVar: string | undefined, value: string | undefined): void {
+function setEnvVar(envVarName: string, value: string | undefined): void {
   if (value) {
-    envVar = value;
+    process.env[envVarName] = value;
   }
 }
 
@@ -53,7 +53,7 @@ export function setEnvEncodeAccessJwtConfig(config: EncodeAccessJwtConfig): void
   process.env.ACCESS_JWT_ALGORITHM = config.algorithm;
   process.env.ACCESS_JWT_ISSUER = config.issuer;
   process.env.ACCESS_JWT_AUDIENCE = config.audience;
-  setEnvVar(process.env.ACCESS_JWT_EXPIRATION, config.expiresIn?.toString());
+  setEnvVar('ACCESS_JWT_EXPIRATION', config.expiresIn?.toString());
 }
 
 export function setEnvEncodeRefreshJwtConfig(config: EncodeRefreshJwtConfig): void {
@@ -61,21 +61,21 @@ export function setEnvEncodeRefreshJwtConfig(config: EncodeRefreshJwtConfig): vo
   process.env.REFRESH_JWT_ALGORITHM = config.algorithm;
   process.env.REFRESH_JWT_ISSUER = config.issuer;
   process.env.REFRESH_JWT_AUDIENCE = config.audience;
-  setEnvVar(process.env.REFRESH_JWT_EXPIRATION, config.expiresIn?.toString());
+  setEnvVar('REFRESH_JWT_EXPIRATION', config.expiresIn?.toString());
 }
 
 export function setEnvDecodeAccessJwtConfig(config: DecodeAccessJwtConfig): void {
   process.env.ACCESS_JWT_PUBLIC_KEY = config.secretOrPublicKey;
   process.env.ACCESS_JWT_AUDIENCE = config.audience;
   process.env.ACCESS_JWT_ISSUER = config.issuer;
-  setEnvVar(process.env.ACCESS_JWT_EXPIRATION, config.expiresIn?.toString());
+  setEnvVar('ACCESS_JWT_EXPIRATION', config.expiresIn?.toString());
 }
 
 export function setEnvDecodeRefreshJwtConfig(config: DecodeRefreshJwtConfig): void {
   process.env.REFRESH_JWT_PUBLIC_KEY = config.secretOrPublicKey;
   process.env.REFRESH_JWT_AUDIENCE = config.audience;
   process.env.REFRESH_JWT_ISSUER = config.issuer;
-  setEnvVar(process.env.REFRESH_JWT_EXPIRATION, config.expiresIn?.toString());
+  setEnvVar('REFRESH_JWT_EXPIRATION', config.expiresIn?.toString());
 }
 
 export function setEnvUserBaseStoreConfig(config: UserBaseStoreConfig): void {